Upgrade the microbit to the latest device framework, and fix the bug that the microbit serial data is not displayed in the terminal at the same time
